update assignment 1
fix bug delete
fix img

+++ Assignment_1_Nguyễn Đình Thọ_HE150527/Ass01Solution/MyStoreWinApp/frmMemberManagement.cs
@@ -117,7 +117,9 @@ namespace MyStoreWinApp
             try
             {
             var member = GetMemberObject();
-            memberRepository.deleteMember(member.MemberID);
+                if (member != null) {
+                    memberRepository.deleteMember(member.MemberID);
+                }
             LoadDVGMemberList();
             }
             catch(Exception ex)
